requirements.txt
fastapi
uvicorn
requests
faiss-cpu
numpy
sentence-transformers
python-dotenv

